Added missing cast of the raw adler to a u32

+++ src/mem.rs
@@ -354,7 +354,7 @@ impl Decompress {
             ffi::MZ_BUF_ERROR => Ok(Status::BufError),
             ffi::MZ_STREAM_END => Ok(Status::StreamEnd),
             ffi::MZ_NEED_DICT => Ok(Status::NeedDictionary { 
-                adler: raw.adler 
+                adler: raw.adler as u32
             }),
             c => panic!("unknown return code: {}", c),
         }
